change BigDecimal for Double in response just for rapidoc purpose

+++ src/main/java/com/mateuszk/controller/PolicyController.java
@@ -32,7 +32,7 @@ public class PolicyController {
     public HttpResponse<PremiumResponse> calculatePolicy(@Body @Valid @NotNull PolicyDTO policyDTO) {
         return HttpResponse
                 .status(HttpStatus.OK)
-                .body(new PremiumResponse(policyDTO.getPolicyNumber(), policyService.calculate(policyDTO)));
+                .body(new PremiumResponse(policyDTO.getPolicyNumber(), policyService.calculate(policyDTO).doubleValue()));
     }
     //@formatter:on
 }
